added 64 bit updates

+++ arch/powerpc/disassembler.cpp
@@ -212,7 +212,7 @@ powerpc_release(void)
 }
 
 extern "C" int
-powerpc_decompose(const uint8_t *data, int size, uint32_t addr, bool lil_end,
+powerpc_decompose(const uint8_t *data, int size, uint64_t addr, bool lil_end,
 	struct decomp_result *res, bool is_64bit, int cs_mode_arg)
 {
 	int rc = -1;
